fix: generating icon type definitions
