+++ packages/core/components/Icon.astro
@@ -1,8 +1,8 @@
 ---
 import icons, { config } from "virtual:astro-icon";
 import { getIconData, iconToSVG } from "@iconify/utils";
-import { cache } from "./cache.mjs";
-// @ts-expect-error - generated on startup
+import { cache } from "./cache";
+// @ts-expect-error - Types are generated by the integration
 import type { Icon } from "astro-icon";
 
 interface Props {
@@ -52,7 +52,7 @@ if (!collection) {
 
 const iconData = getIconData(collection, iconName ?? setName);
 if (!iconData) {
-  const err: any = new Error(`Unable to locate "${name}" icon!`);
+  const err = new Error(`Unable to locate "${name}" icon!`);
 
   if (import.meta.env.DEV) {
     const { include = {} } = config;
