+++ packages/core/components/index.ts
@@ -1,0 +1,3 @@
+/// <reference types="astro/astro-jsx" />
+
+export { default as Icon } from './Icon.astro'
