+++ packages/core/package.json
@@ -2,19 +2,17 @@
   "name": "astro-icon",
   "version": "0.0.0",
   "type": "module",
+  "types": "./dist/integration.d.ts",
   "exports": {
     ".": {
       "types": "./dist/integration.d.ts",
-      "default": "./dist/integration.mjs"
+      "default": "./dist/integration.js"
     },
-    "./components": {
-      "types": "./dist/components.d.mts",
-      "default": "./dist/components.mjs"
-    }
+    "./components": "./components/index.js"
   },
   "files": [
-    "dist",
-    "lib"
+    "components",
+    "dist"
   ],
   "scripts": {
     "build": "tsc",
@@ -47,9 +45,9 @@
   },
   "homepage": "https://github.com/natemoo-re/astro-icon#readme",
   "dependencies": {
-    "@iconify/tools": "^2.2.6",
+    "@iconify/tools": "^3.0.1",
     "@iconify/types": "^2.0.0",
-    "@iconify/utils": "^2.1.4"
+    "@iconify/utils": "^2.1.5"
   },
   "devDependencies": {
     "@astrojs/ts-plugin": "^1.0.6",
