--- packages/core/src/components.mts
@@ -1,2 +1,0 @@
-// @ts-ignore - TODO: Fix error
-export { default as Icon } from '../lib/Icon.astro'
