+++ packages/core/src/loaders/index.ts
@@ -1,0 +1,2 @@
+export { default as loadLocalCollection} from './loadLocalCollection'
+export { default as loadIconifyCollections, loadCollection } from './loadIconifyCollections'
