+++ packages/core/src/loaders/loadLocalCollection.ts
@@ -6,8 +6,8 @@ import {
   parseColors,
   runSVGO,
 } from "@iconify/tools";
-import type { IconCollection } from "../typings/integrationOptions.d.ts";
-import type { SVG, Color, SVGOOptions } from '../typings/iconify.d.ts'
+import type { IconCollection } from "virtual:astro-icon";
+import type { SVG, Color, SVGOOptions } from '../../typings/iconify.js'
 
 
 export default async function createLocalCollection(dir: string, options?: SVGOOptions): Promise<IconCollection> {
@@ -33,7 +33,7 @@ export default async function createLocalCollection(dir: string, options?: SVGOO
     // Clean up and optimize icons
     try {
       // Clean up icon code
-      await cleanupSVG(svg);
+      cleanupSVG(svg);
 
       if (await isMonochrome(svg)) {
         await convertToCurrentColor(svg);
