--- packages/core/typings/integrationOptions.d.ts
@@ -1,13 +1,0 @@
-import type { IconifyJSON } from "./iconify";
-
-export type IntegrationOptions = {
-  include?: Record<string, ['*'] | string[]>
-  /**
-   * @default "src/icons"
-   */
-  iconDir?: string
-};
-
-
-export type IconCollection = IconifyJSON;
-export type AstroIconCollectionMap = Record<string, IconCollection>;
