+++ packages/core/typings/virtual.d.ts
@@ -1,6 +1,10 @@
 declare module "virtual:astro-icon" {
-  import type { IntegrationOptions, AstroIconCollectionMap } from './integrationOptions'
-  
+  import type { IntegrationOptions } from "../src/integration";
+  import type { IconifyJSON } from "./iconify";
+
+  export type IconCollection = IconifyJSON;
+  export type AstroIconCollectionMap = Record<string, IconCollection>;
+
   const icons: AstroIconCollectionMap;
   export default icons;
   export const config: IntegrationOptions;
