+++ packages/www/astro.config.mjs
@@ -1,3 +1,4 @@
+// @ts-check
 import { defineConfig } from 'astro/config';
 import icon from 'astro-icon';
